backend/app/core/jwt.py
from datetime import datetime, timedelta
from jose import jwt, JWTError

SECRET_KEY = "I6tQqKQOjVK1/B1BCJOWTyuxi5VMu+fRQ/t6OoApOKuVWkjzNHwzyKLg/YVi1w+M"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_access_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
